- All provided image derivative filters now can compute the derivative along the image's edge - Improved unit testing for image derivatives - Added sobel based hessian - Added factory for abstracted convolutions - Added generalized image filter interface - Added/updated several unit test - Restructed benchmarks for image derivatives - Added benchmark for extended convolutions

+++ core/src/gecv/alg/detect/corner/impl/KltCorner_I16.java
@@ -36,8 +36,8 @@ public class KltCorner_I16 extends SsdCorner_I16 implements KltCornerIntensity<I
 	@Override
 	protected float computeIntensity() {
 		// compute the smallest eigenvalue
-		float left = (totalXX + totalYY) * 0.5f;
-		float b = (totalXX - totalYY) * 0.5f;
+		double left = (totalXX + totalYY) * 0.5f;
+		double b = (totalXX - totalYY) * 0.5f;
 		double right = Math.sqrt(b * b + (double)totalXY * totalXY);
 
 		// the smallest eigenvalue will be minus the right side
